// FE/src/app/Shared/models/Address.model.ts
export interface District {
  name: string;
  code: number;
  division_type: string;
  codename: string;
  province_code: number;
  wards: Ward[];
}

export interface Province {
  name: string;
  code: number;
  division_type: string;
  codename: string;
  phone_code: number;
  districts: District[];
}

export interface Ward {
  name: string;
  code: number;
  division_type: string;
  codename: string;
  phone_code: number;
}